+++ examples/nextjs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @davstack/example-1
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @davstack/sound@0.1.1
+
 ## 0.0.1
 
 ### Patch Changes